Some refactoring work. I've adding new projects instead of messing with existing ones to keep them there for further reference and because i usually code at home where i don't have an internet connection and i can't revert changes so i need the original copies :S
I've also added a Extras solution to put there all the subprojects with usefull implementations that can be useful in another projects.

I've also splitted the SharpKnocking.Common assembly in another two to keep the native methods only in one assembly and the rest in another. This will help if we are in the need of having an assembly with the native methods for each concrete os.